Base code rebuilt.
In this commit the base code has been rebuilt to provide clean skeleton.
All game components remain intact including screen, basic windows and other.
The game and editor gadgets and engine will be adapted to the new scheme.
The function protocols have been set.
