+++ .gitignore
@@ -5,3 +5,4 @@
 *.lev
 *.zip
 Fonts
+Docs/build
