+++ Headers/Game.h
@@ -1,0 +1,6 @@
+
+/* Functions that work on GameData */
+
+/* Default GameData action */
+
+long resetGame( struct gameData *gd, struct programData *pd );
