+++ Headers/GameData.h
@@ -1,0 +1,7 @@
+
+#include "BoardData.h"
+
+struct gameData
+{
+    struct boardData boardData;
+};
