+++ Headers/Program.h
@@ -1,0 +1,4 @@
+
+/* Functions that work on ProgramData */
+
+long resetProgram( struct programData *pd );
