+++ Headers/System.h
@@ -1,0 +1,2 @@
+
+long resetSystem(struct systemData *sd, struct programData *pd);
