+++ Headers/SystemData.h
@@ -1,0 +1,7 @@
+
+/* Operating system data */
+
+struct systemData
+{
+    int i;
+};
