+++ Headers/Tile.h
@@ -1,53 +1,8 @@
 
-/*
-** Silnik C-1200
-*/
+/* Functions that work on TileData */
 
-/* Pole na planszy gry Magazyn */
+long replaceFloor(int floorType, int floorID, struct floorData *fd, struct tileData *td, struct boardData *bd, struct gameData *gd, struct programData *pd);
 
-/* Udostêpnione elementy */
+long replaceObject(int objectType, int objectID, struct objectData *od, struct tileData *td, struct boardData *bd, struct gameData *gd, struct programData *pd);
 
-#include <exec/types.h>
-
-/* Domy¶lny rozmiar */
-
-#define COLS 20
-#define ROWS 16
-
-enum /* Podstawowe kafle */
-{
-    BACK,
-    WALL,
-    FLOOR,
-    FLAGSTONE
-    BASIC_TILES
-};
-
-enum /* Podstawowe obiekty */
-{
-    NONE,
-    HERO,
-    BOX,
-    BASIC_OBJECTS
-};
-
-struct TileLayer
-{
-    UBYTE type, id;
-};
-
-struct Tile
-{
-    struct TileLayer floor, object;
-};
-
-extern struct TileAPI
-{
-    void (*init)(); /* Przygotuj API do pracy */
-    void (*clear)(struct Tile *t); /* Wyczy¶æ tablicê kafli */
-    struct Tile *(*alloc)(); /* Zaalokuj planszê z³o¿on± z kafli */
-    
-
-
-
-} tile;
+long tile(struct tileData *td, struct boardData *bd, struct gameData *gd, struct programData *pd);
