+++ Src/System/System.c
@@ -1,0 +1,10 @@
+
+#include "SystemData.h"
+#include "System.h"
+
+long resetSystem(struct systemData *sd, struct programData *pd)
+{
+    /* Reset system components (libs, devs, screen etc.) */
+
+
+}
